backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-nmap==0.7.1
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
pydantic==2.5.0
pydantic-settings==2.1.0
redis==5.0.1
celery==5.3.4
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dateutil==2.8.2
asyncpg==0.29.0
aiofiles==23.2.1
httpx==0.25.2
